chore: Timer를 contracts에서 utils로 이동, 약간의 주석 추가

+++ src/tetris/tetromino.rs
@@ -18,7 +18,7 @@ impl Block for I {
     }
 
     fn rotate(&mut self, n: u8) {
-        self.rotate_index.add(n as i8);
+        self.rotate_index.add(n);
     }
 
     fn move_left(&mut self) {
